backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6
google-generativeai==0.3.1
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.1
numpy==1.26.2
scipy==1.11.4
